changed bigint to long

+++ src/main/java/com/napier/sem/Population.java
@@ -5,8 +5,7 @@ import java.math.BigInteger;
  */
 public class Population {
     public String reportName;
-    public BigInteger totalPopulation;
-    // public double totalPopulation;
+    public long totalPopulation;
     public int totalCity;
     public int totalNotCity;
